rename profile context arrange the code swap controllers
