+++ api/HotDeskApplicationApi/Data/HotDeskDbContext.cs
@@ -4,20 +4,16 @@ using Microsoft.EntityFrameworkCore;
 
 namespace HotDeskApplicationApi.Data
 {
-	public class ProfileContext : IdentityDbContext
+	public class HotDeskDbContext : IdentityDbContext
 	{
         public DbSet<Models.Profile> Profile { get; set; } = default!;
 
-        public ProfileContext(DbContextOptions<ProfileContext> options)
+        public HotDeskDbContext(DbContextOptions<HotDeskDbContext> options)
             : base(options)
         { }
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
             base.OnModelCreating(modelBuilder);
-
-            modelBuilder
-                .Entity<Models.Profile>()
-                .HasKey(nameof(Models.Profile.ID));
         }
     }
 }
