+++ api/HotDeskApplicationApi/Models/Profile.cs
@@ -4,9 +4,12 @@ namespace HotDeskApplicationApi.Models
 	public class Profile
 	{
         public Guid ID { get; set; }
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
-        public string EmailAddress { get; set; }
+
+        public string FirstName { get; set; } = default!;
+
+        public string LastName { get; set; } = default!;
+
+        public string EmailAddress { get; set; } = default!;
     }
 }
