+++ api/HotDeskApplicationApi/Models/Security/LoginModel.cs
@@ -3,9 +3,9 @@ namespace HotDeskApplicationApi.Models.Security
 {
 	public class LoginModel
 	{
-        public string Email { get; set; }
+        public string Email { get; set; } = default!;
 
-        public string Password { get; set; }
+        public string Password { get; set; } = default!;
     }
 }
