+++ api/HotDeskApplicationApi/Models/Security/RegisterModel.cs
@@ -3,13 +3,13 @@ namespace HotDeskApplicationApi.Models.Security
 {
 	public class RegisterModel
 	{
-        public string FirstName { get; set; }
+        public string FirstName { get; set; } = default!;
 
-        public string LastName { get; set; }
+        public string LastName { get; set; } = default!;
 
-        public string Email { get; set; }
+        public string Email { get; set; } = default!;
 
-        public string Password { get; set; }
+        public string Password { get; set; } = default!;
     }
 }
