+++ api/HotDeskApplicationApi/Models/Security/Token.cs
@@ -3,7 +3,7 @@ namespace HotDeskApplicationApi.Models.Security
 {
 	public class Token
 	{
-        public string Value { get; set; }
+        public string Value { get; set; } = default!;
 
         public DateTime Expiry { get; set; }
     }
