+++ api/HotDeskApplicationApi/appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "ProfileContext": "Host=localhost; Database=HotDeskApp; Username=postgres; Password=cristina123"
+    "HotDeskDbContext": "Host=localhost; Database=HotDeskApp; Username=postgres; Password=cristina123"
   },
   "Logging": {
     "LogLevel": {
